create a chat Interactor instead of chat viewmodel
